03-app-clima/weather-backend/requirements.txt
Flask==2.1.2
requests==2.26.0
werkzeug==2.1.2
flask_cors==2.1.2
python-dotenv==0.19.2


